#8 Added Dictionary, Adjust Context Menu
Dictionary for adding Keywords and their Definition. Json is always the same whatever the lds file opened so rest assured that the dictionary will be there. Removed Solution on the log type selection and will appear on Problem context menu. This will help balance the # and make the two directly proportional. Fix is now the word for bug context menu