#10 Filter, Welcome Frame, Setup Frame
Filter logs is added. Welcome screen is now the main screen to be display because the project is scaling into new concepts. LogApp breaks into two debugging and general basically the same window but different functionality. Debugging only allow log types: Just Details and Bugs and also the .lds file is not the default saving extension anymore, thus creating new .ldsg for general, .ldsd for debugging and many more upcoming extension with appending log mode on the end of .lds. Currently working on the setup page